sample type and data class design/data export options
