+++ api/src/org/labkey/api/admin/FolderArchiveDataTypes.java
@@ -45,6 +45,10 @@ public class FolderArchiveDataTypes
     public static final String FILE_BROWSER_SETTINGS = "File browser admin configurations";
     public static final String ETLS = "ETL Definitions";
     public static final String SAMPLE_TYPES_AND_DATA_CLASSES = "Sample Types and Data Classes";
+    public static final String SAMPLE_TYPE_DESIGNS = "Sample Type Designs";
+    public static final String SAMPLE_TYPE_DATA = "Sample Type Data";
+    public static final String DATA_CLASS_DESIGNS = "Data Class Designs";
+    public static final String DATA_CLASS_DATA = "Data Class Data";
     public static final String INVENTORY = "Inventory locations and items";
     public static final String VIEW_CATEGORIES = "Categories";
 }
