+++ experiment/src/org/labkey/experiment/samples/AbstractExpFolderImporter.java
@@ -2,10 +2,8 @@ package org.labkey.experiment.samples;
 
 import org.apache.logging.log4j.Logger;
 import org.jetbrains.annotations.Nullable;
-import org.labkey.api.admin.FolderArchiveDataTypes;
 import org.labkey.api.admin.FolderImportContext;
 import org.labkey.api.admin.FolderImporter;
-import org.labkey.api.admin.FolderImporterFactory;
 import org.labkey.api.data.Container;
 import org.labkey.api.data.DbScope;
 import org.labkey.api.data.TableInfo;
@@ -13,20 +11,15 @@ import org.labkey.api.dataiterator.DataIteratorContext;
 import org.labkey.api.dataiterator.DetailedAuditLogDataIterator;
 import org.labkey.api.exp.ExperimentException;
 import org.labkey.api.exp.FileXarSource;
-import org.labkey.api.exp.Identifiable;
 import org.labkey.api.exp.XarSource;
 import org.labkey.api.exp.api.ExpSampleType;
 import org.labkey.api.exp.api.ExperimentService;
 import org.labkey.api.exp.api.SampleTypeService;
-import org.labkey.api.exp.query.ExpSchema;
-import org.labkey.api.exp.query.SamplesSchema;
 import org.labkey.api.pipeline.PipelineJob;
 import org.labkey.api.query.AbstractQueryUpdateService;
 import org.labkey.api.query.BatchValidationException;
-import org.labkey.api.query.InvalidKeyException;
 import org.labkey.api.query.QueryService;
 import org.labkey.api.query.QueryUpdateService;
-import org.labkey.api.query.QueryUpdateServiceException;
 import org.labkey.api.query.UserSchema;
 import org.labkey.api.query.ValidationException;
 import org.labkey.api.reader.DataLoader;
@@ -44,43 +37,20 @@ import java.io.InputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.sql.SQLException;
-import java.util.Comparator;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
 
 import static org.labkey.api.admin.FolderImportContext.IS_NEW_FOLDER_IMPORT_KEY;
-import static org.labkey.experiment.samples.SampleTypeAndDataClassFolderWriter.DEFAULT_DIRECTORY;
-import static org.labkey.experiment.samples.SampleTypeAndDataClassFolderWriter.EXCLUDED_TYPES;
-import static org.labkey.experiment.samples.SampleTypeAndDataClassFolderWriter.XAR_RUNS_NAME;
-import static org.labkey.experiment.samples.SampleTypeAndDataClassFolderWriter.XAR_RUNS_XML_NAME;
-import static org.labkey.experiment.samples.SampleTypeAndDataClassFolderWriter.XAR_TYPES_NAME;
-import static org.labkey.experiment.samples.SampleTypeAndDataClassFolderWriter.XAR_TYPES_XML_NAME;
+import static org.labkey.experiment.samples.AbstractExpFolderWriter.XAR_RUNS_NAME;
+import static org.labkey.experiment.samples.AbstractExpFolderWriter.XAR_RUNS_XML_NAME;
 
-public class SampleTypeAndDataClassFolderImporter implements FolderImporter
+public abstract class AbstractExpFolderImporter implements FolderImporter
 {
-    // registry data classes have to be imported in a particular order because they reference each other
-    private static final List<String> REGISTRY_CLASS_ORDER = List.of(
-            "ProtSequence",
-            "NucSequence",
-            "Molecule",
-            "Vector",
-            "Construct",
-            "CellLine",
-            "ExpressionSystem",
-            "Compound"
-    );
-
-    protected SampleTypeAndDataClassFolderImporter()
-    {
-    }
-
-    @Override
-    public String getDataType()
-    {
-        return FolderArchiveDataTypes.SAMPLE_TYPES_AND_DATA_CLASSES;
-    }
+    protected abstract VirtualFile getXarDir(VirtualFile root);
+    protected abstract void importDataFiles(FolderImportContext ctx, VirtualFile root, XarReader typesReader) throws IOException, SQLException;
+    protected abstract boolean isXarTypesFile(String fileName);
+    protected abstract boolean excludeTable(String tableName);
 
     @Override
     public String getDescription()
@@ -91,7 +61,7 @@ public class SampleTypeAndDataClassFolderImporter implements FolderImporter
     @Override
     public void process(@Nullable PipelineJob job, FolderImportContext ctx, VirtualFile root) throws Exception
     {
-        VirtualFile xarDir = root.getDir(DEFAULT_DIRECTORY);
+        VirtualFile xarDir = getXarDir(root);
 
         if (xarDir != null)
         {
@@ -99,17 +69,15 @@ public class SampleTypeAndDataClassFolderImporter implements FolderImporter
             Path xarDirPath = Path.of(xarDir.getLocation());
             Path typesXarFile = null;
             Path runsXarFile = null;
-            Map<String, String> sampleTypeDataFiles = new HashMap<>();
-            Map<String, String> dataClassDataFiles = new HashMap<>();
             Logger log = ctx.getLogger();
 
             if (null != job)
                 job.setStatus("IMPORT " + getDescription());
-            log.info("Starting Sample Type and Data Class import");
+            log.info("Starting " + getDescription());
 
             for (String file: xarDir.list())
             {
-                if (file.equalsIgnoreCase(XAR_TYPES_NAME) || file.equalsIgnoreCase(XAR_TYPES_XML_NAME))
+                if (isXarTypesFile(file))
                 {
                     if (typesXarFile == null)
                         typesXarFile = xarDirPath.resolve(file);
@@ -122,17 +90,6 @@ public class SampleTypeAndDataClassFolderImporter implements FolderImporter
                         runsXarFile = xarDirPath.resolve(file);
                     else
                         log.error("More than one runs XAR file found in the sample type directory: ", file);
-                }
-                else if (file.toLowerCase().endsWith(".tsv"))
-                {
-                    if (file.startsWith(SampleTypeAndDataClassFolderWriter.SAMPLE_TYPE_PREFIX))
-                    {
-                        sampleTypeDataFiles.put(FileUtil.getBaseName(file.substring(SampleTypeAndDataClassFolderWriter.SAMPLE_TYPE_PREFIX.length())), file);
-                    }
-                    else if (file.startsWith(SampleTypeAndDataClassFolderWriter.DATA_CLASS_PREFIX))
-                    {
-                        dataClassDataFiles.put(FileUtil.getBaseName(file.substring(SampleTypeAndDataClassFolderWriter.DATA_CLASS_PREFIX.length())), file);
-                    }
                 }
             }
 
@@ -153,19 +110,8 @@ public class SampleTypeAndDataClassFolderImporter implements FolderImporter
                     typesReader.setStrictValidateExistingSampleType(xarCtx.isStrictValidateExistingSampleType());
                     typesReader.parseAndLoad(false, ctx.getAuditBehaviorType());
 
-                    // Import data classes first because the media sample type (RawMaterials) has a lookup to the ingredient data class.
-                    // Registry files need to imported in a particular order because some files rely on data from other files.
-                    if (!dataClassDataFiles.isEmpty())
-                    {
-                        importTsvData(ctx, ExpSchema.SCHEMA_EXP_DATA.toString(), typesReader.getDataClasses().stream().map(Identifiable::getName).sorted(Comparator.comparing(REGISTRY_CLASS_ORDER::indexOf)).collect(Collectors.toList()),
-                                dataClassDataFiles, xarDir, true, false);
-                    }
-
-                    if (!sampleTypeDataFiles.isEmpty())
-                    {
-                        importTsvData(ctx, SamplesSchema.SCHEMA_NAME, typesReader.getSampleTypes().stream().map(Identifiable::getName).collect(Collectors.toList()),
-                                sampleTypeDataFiles, xarDir, true, false);
-                    }
+                    // import any exported tsv data files
+                    importDataFiles(ctx, xarDir, typesReader);
 
                     // handle wiring up any derivation runs
                     if (runsXarFile != null)
@@ -189,7 +135,7 @@ public class SampleTypeAndDataClassFolderImporter implements FolderImporter
                         runsReader.setStrictValidateExistingSampleType(xarCtx.isStrictValidateExistingSampleType());
                         runsReader.parseAndLoad(false, ctx.getAuditBehaviorType());
                     }
-                    // handle aliquot rollup calculation for all sample types
+                    // handle aliquot rollup calculation for all sample types, this is a noop for the data class importer
                     for (ExpSampleType sampleType : typesReader.getSampleTypes())
                         SampleTypeService.get().recomputeSampleTypeRollup(sampleType, ctx.getContainer());
                 }
@@ -197,14 +143,14 @@ public class SampleTypeAndDataClassFolderImporter implements FolderImporter
                     log.info("No types XAR file to process.");
 
                 transaction.commit();
-                log.info("Finished importing Sample Types and Data Classes");
+                log.info("Finished " + getDescription());
             }
         }
     }
 
     protected XarReader getXarReader(@Nullable PipelineJob job, FolderImportContext ctx, VirtualFile root, Path typesXarFile) throws IOException, ExperimentException
     {
-        VirtualFile xarDir = root.getDir(DEFAULT_DIRECTORY);
+        VirtualFile xarDir = getXarDir(root);
         Logger log = ctx.getLogger();
 
         Path logFile = null;
@@ -272,7 +218,7 @@ public class SampleTypeAndDataClassFolderImporter implements FolderImporter
         };
     }
 
-    protected void importTsvData(FolderImportContext ctx, String schemaName, List<String> tableNames, Map<String, String> dataFileMap, VirtualFile dir, boolean fileRequired, boolean isUpdate) throws IOException, SQLException, BatchValidationException, QueryUpdateServiceException, InvalidKeyException
+    protected void importTsvData(FolderImportContext ctx, String schemaName, List<String> tableNames, Map<String, String> dataFileMap, VirtualFile dir, boolean fileRequired, boolean isUpdate) throws IOException, SQLException
     {
         Logger log = ctx.getLogger();
         UserSchema userSchema = QueryService.get().getUserSchema(ctx.getUser(), ctx.getContainer(), schemaName);
@@ -346,7 +292,7 @@ public class SampleTypeAndDataClassFolderImporter implements FolderImporter
                         log.error("Failed to find table '" + schemaName + "." + tableName + "' to import data file: " + dataFileName);
                     }
                 }
-                else if (fileRequired && !EXCLUDED_TYPES.contains(tableName))
+                else if (fileRequired && !excludeTable(tableName))
                 {
                     log.error("Unable to import TSV data for table " + tableName + ". File not found.");
                 }
@@ -355,22 +301,6 @@ public class SampleTypeAndDataClassFolderImporter implements FolderImporter
         else
         {
             log.error("Could not find " + schemaName + " schema.");
-        }
-    }
-
-    public static class Factory implements FolderImporterFactory
-    {
-        @Override
-        public FolderImporter create()
-        {
-            return new SampleTypeAndDataClassFolderImporter();
-        }
-
-        @Override
-        public int getPriority()
-        {
-            // make sure this importer runs before the FolderXarImporter (i.e. "Experiments and runs")
-            return 65;
         }
     }
 }
